# back/api/src/public/docs/openapi.yaml
openapi: 3.0.0
info:
  title: 'Пример реализации социальной сети'
  description: "## Соцальная сеть Шаг и компания \n\n Наша проба сделать хорошую социальную сеть"
  version: 1.0.0
servers:
  -
    url: 'http://localhost'
    description: 'Local API server'
paths:
...
  /api/account/update-user:
    put:
      tags:
        - account
      summary: 'Update user data (auth Bearer)'
      operationId: api.account.update.user
      requestBody:
        description: 'update data is not required some'
        required: true
        content:
          application/json:
            schema:
              properties:
                user_id:
                  type: integer
                  example: '22'
                name:
                  type: string
                  example: REcountsx
                password:
                  type: password
                  example: qwerty123
              type: object
      responses:
        '200':
          description: 'Admin his updated'
          content:
            application/json:
              schema:
                properties:
                  status: { type: string, example: 'Admin his updated' }
                type: object
...
  /api/auth/register:
    post:
      tags:
        - auth
      summary: 'Register a user'
      operationId: api.auth.register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessLoginResource'
        '422':
          description: 'password confirmation'
          content:
            application/json:
              schema:
                properties:
                  status: { type: string, example: error }
                  message: { type: string, example: 'password confirmation.' }
                type: object
...
components:
  schemas:
...
    RegisterRequest:
      required:
        - name
        - email
        - password
        - password_confirmation
      properties:
        name:
          type: string
          example: 'John Doe'
        email:
          type: string
          format: email
          example: john@example.com
        password:
          type: string
          example: '123456'
        password_confirmation:
          type: string
          example: '123456'
      type: object
...
    SuccessLoginResource:
      properties:
        status:
          type: string
          example: 'true'
        user:
          $ref: '#/components/schemas/Admin'
        authorisation:
          $ref: '#/components/schemas/TokenResource'
      type: object
    TokenResource:
      properties:
        token:
          type: string
          example: eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwOi8vbG9jYWxob3N0L2FwaS9hdXRoL3JlZ2lzdGVyIiwiaWF0IjoxNzA2NTUyMTM4LCJleHAiOjE3MDY5MTIxMzgsIm5iZiI6MTcwNjU1MjEzOCwianRpIjoiRk9hclkyZ2VvckhPT1llbSIsInN1YiI6IjQiLCJwcnYiOiIyM2JkNWM4OTQ5ZjYwMGFkYjM5ZTcwMWM0MDA4NzJkYjdhNTk3NmY3IiwiZW1haWwiOiJ2YXN5YTJAdmFzeWEuY29tIiwibmFtZSI6InZhc3lhMiJ9.APeVTaT7gnkVVboKkfXDndq1vjQFiiaMPCw21aS9hL8
        type:
          type: string
          example: bearer
      type: object
...
    Admin:
      properties:
        id:
          type: integer
          example: '2'
        name:
          type: string
          example: 'John Doe'
        email:
          type: string
          format: email
          example: john@example.com
        email_verified:
          type: boolean
          example: '1'
        created_at:
          type: string
          format: date-time
          example: '2024-02-03T21:06:04.000000Z'
        updated_at:
          type: string
          format: date-time
          example: '2024-02-03T21:06:04.000000Z'
        confirmation_token:
          type: url
          example: 8ccde004-4c26-453e-a3f7-a73145dfb4a1
        avatar_url:
          type: string
          example: 'http://localhost/storage/avatars/108/original.webp'
      type: object
  securitySchemes:
    BearerToken:
      type: http
      bearerFormat: JWT
      scheme: bearer
...